Move shared layout components
