+++ src/shared/Footer.tsx
@@ -3,9 +3,7 @@ import "./Footer.css";
 function Footer() {
   return (
     <footer className="footer">
-      <p>
-        © {new Date().getFullYear()} Sylvester Kitaka Munyao
-      </p>
+      <p>© {new Date().getFullYear()} Sylvester Kitaka Munyao</p>
     </footer>
   );
 }
